Splitted layout integrations.
